First draft of the README and code updates
I wrote a first draft of the README.adoc file, to walk through the basics of the code and point out that this version is annotation-based. I also changed the names of the objects and paths in the code to get everything to match. I also pulled code that I had repeated in three places to a helper function, even though that code is central to the operation at hand. On reflection, having it in a method of its own may help to highlight it.